Fix three x64 crashes: SCNotification layout, import button reentrancy, GetOpenFileName
- SCNotification.position (and length/linesAdded/line/annotationLinesAdded) were
  declared as int in old Scintilla.h, but modern Notepad++ 64-bit sends them as
  Sci_Position (intptr_t, 8 bytes). The layout mismatch shifted the text pointer
  field by 4-8 bytes, causing an access violation when beNotified read
  notifyCode->text during SCN_AUTOCSELECTION (hotspot list navigation).
  Fix: widen the five fields to intptr_t to match modern Scintilla 5.x layout.

- importSnippets: move g_freezeDock=true/LIVE_HINT_UPDATE-- to the top so no
  NPPN_BUFFERACTIVATED can re-enter updateDockItems before the freeze is set.
  Fix IDCANCEL early-return that skipped the freeze/counter cleanup.
  Fix GetSaveFileName -> GetOpenFileName (was showing Save-As dialog for import).

- exportSnippets: save/restore g_freezeDock instead of unconditionally clearing
  it, so backupAllSnippets called from importSnippets does not undo the freeze.

- Dock import button: call deferImportSnippets() (PostMessage) instead of
  importSnippetsOnly() directly, so the import runs from NPP's main message loop
  rather than synchronously inside the dock's WM_COMMAND handler, preventing
  reentrancy from the modal file dialog and NPP_MENUCOMMAND calls.

+++ NppApi/Scintilla.h
@@ -1117,7 +1117,7 @@ struct Sci_NotifyHeader {
 
 struct SCNotification {
 	struct Sci_NotifyHeader nmhdr;
-	int position;
+	intptr_t position;
 	/* SCN_STYLENEEDED, SCN_DOUBLECLICK, SCN_MODIFIED, SCN_MARGINCLICK, */
 	/* SCN_NEEDSHOWN, SCN_DWELLSTART, SCN_DWELLEND, SCN_CALLTIPCLICK, */
 	/* SCN_HOTSPOTCLICK, SCN_HOTSPOTDOUBLECLICK, SCN_HOTSPOTRELEASECLICK, */
@@ -1133,12 +1133,12 @@ struct SCNotification {
 	const char *text;
 	/* SCN_MODIFIED, SCN_USERLISTSELECTION, SCN_AUTOCSELECTION, SCN_URIDROPPED */
 
-	int length;		/* SCN_MODIFIED */
-	int linesAdded;	/* SCN_MODIFIED */
+	intptr_t length;	/* SCN_MODIFIED */
+	intptr_t linesAdded;	/* SCN_MODIFIED */
 	int message;	/* SCN_MACRORECORD */
 	uptr_t wParam;	/* SCN_MACRORECORD */
 	sptr_t lParam;	/* SCN_MACRORECORD */
-	int line;		/* SCN_MODIFIED */
+	intptr_t line;		/* SCN_MODIFIED */
 	int foldLevelNow;	/* SCN_MODIFIED */
 	int foldLevelPrev;	/* SCN_MODIFIED */
 	int margin;		/* SCN_MARGINCLICK */
@@ -1146,7 +1146,7 @@ struct SCNotification {
 	int x;			/* SCN_DWELLSTART, SCN_DWELLEND */
 	int y;		/* SCN_DWELLSTART, SCN_DWELLEND */
 	int token;		/* SCN_MODIFIED with SC_MOD_CONTAINER */
-	int annotationLinesAdded;	/* SCN_MODIFIED with SC_MOD_CHANGEANNOTATION */
+	intptr_t annotationLinesAdded;	/* SCN_MODIFIED with SC_MOD_CHANGEANNOTATION */
 	int updated;	/* SCN_UPDATEUI */
 };
